### 2022-05-26 #4425 @editorialbot
[PRE REVIEW]: `pygetpapers`: a Python library for automated retrieval of scientific literature
**Submitting author:** @ayush4921 (<a href="http://orcid.org/0000-0001-7016-747X">Ayush Garg</a>)
**Repository:** https://github.com/petermr/pygetpapers/
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.3
