### 2022-05-30 #4425 @arfon
:wave: @chartgerink – this looks like a good fit for your domain. I've not done a scope review here but if you are able to edit this, I'd appreciate if you could check if you believe this submission is in scope for JOSS.
