### 2022-05-31 #4425 @chartgerink
@arfon I co-authored a paper with one of the authors in 2017 - which I think falls outside of [the conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) scope?

I want to double check with you - I'd be happy to edit this if you think it appropriate 😊
