### 2022-05-31 #4425 @danielskatz
Yes, this is outside the COI time frame.  Please go ahead and take it if you feel able to do so.
