+++ joss/paper.bib
@@ -11,7 +11,7 @@
   author = {Wind, L.L. and Briganti, J.S. and Brown, A.M.},
   title = {Finding What Is Inaccessible: Antimicrobial Resistance Language Use among the One Health Domains. Antibiotics},
   date = {2021-04-10},
-  doi = {10.3390/antibiotics10040385.},
+  doi = {10.3390/antibiotics10040385},
   note = {PMID: 33916878; PMCID: PMC8065768.},
   pmid = {33916878},
   pmcid = {8065768},
