### 2022-06-08 4ad3b6e
Create CONTRIBUTING.md
