+++ docs/CONTRIBUTING.md
@@ -1,0 +1,5 @@
+Contributions are welcome through issues as well as pull requests. For direct contributions, you can mail the author at ayush@science.org.in.
+
+To discuss problems or feature requests, file an issue. For bugs, please include as much information as possible, including operating system, python version, and version of all dependencies.
+
+To contribute, make a pull request. Contributions should include tests for any new features/bug fixes and follow best practices including PEP8, etc.
