### 2022-06-26 #4451 @khinsen
@chartgerink I have checked all the boxes of my checklist, the authors have replied and reacted quickly to all my comments and issues, so I'd say my review is done! I have to add that "Have the functional claims of the software been confirmed?" is not something I can do exhaustively for a software that has this much diverse functionality. I have run a few tests, which all worked out fine, but I can't claim to have tried all the options of the software, nor all the paper repositories it supports.

Overall, `pygetpapers` is a useful piece of software for anyone doing content mining or just meta-analyses. For geekier scientists, it could even be a tool for bibliography maintenance in the course of a research project.

A challenge for the future would be tutorials that teach these use cases of `pygetpapers` to newcomers. In particular content mining is one of the big added values of the internet age, and not yet as widely practiced as it should in my opinion.
