### 2022-06-30 #4451 @kjappelbaum
I'm also done with the review - while there are minor things one could further improve (e.g. making the tutorials slightly less overwhelming, completing some of the empty docstrings) - it looks generally good to me: the software works, is documented, there is some automated testing, and community guidelines.
