### 2022-07-01 #4451 @chartgerink
@ayush4921 all three reviewers are positive about the paper 👍

I am favorable to move this paper forward for publication, but before being able to do so I will require you to go through [the generated PDF](https://github.com/openjournals/joss-papers/blob/joss.04451/joss.04451/10.21105.joss.04451.pdf) and do an overall proofcheck.

While reading through the paper I noticed several issues that would need to be addressed:

1. Bullet lists or enumerated lists are not displayed as lists - please fix
  - p3
    - lines 92-101
    - lines 112-113
  - p4
    - lines 159 - p5 line 171
  - p5
    - lines 203-205
  - p6
    - lines 203-205
  - p7
    - lines 207-210
    - lines 215-219
3. There are incomplete sentences such as the following - please proofread for reader clarity upon publication and here some of the issues I came across
  - p1
    - "wrap it `tkinter"
  - p2
    - "percountry"
    - "See Data some"
  - p5
    - "`FIRST_PDATE`" seems to miss a letter
4. The use of the word `master` seems unnecessary, please use `main`
5. Please provide a bit more of an explanation of what a `CProject` and a `CTree` are - I am familiar with them and this was still somewhat confusing. It would help educate the reader for use of the software if these are explained a bit more in depth.
6. The CRediT statement does not provide work based on the CRediT taxonomy. Either replace with "Contribution statement" or something of that sort, or specify contributions as based on the CRediT taxonomy.
7. Please add the references header after the credit statement section

Please make the necessary changes and any additional ones you might discover. During this process you can use `@editorialbot generate pdf` to verify your changes are effective.

Once you've made the changes, ping me and I'll take another look 😊 👍
