+++ joss/paper.bib
@@ -53,7 +53,7 @@
     title = {pyami - Semantic Reader of the Scientific Literature},
     version = {0.0.17},
     url = {https://github.com/petermr/pyami/},
-    author = {Peter Murray Rust and others},
+    author = {Peter Murray-Rust and others},
     year = {2021}
 }
 
@@ -69,7 +69,7 @@
     title = {pyamiimage - tools to extract semantic information from scientific diagrams},
     version = {0.0.8},
     url = {https://github.com/petermr/pyamiimage},
-    author = {Peter Murray Rust, Anubhab Chakraborty and others},
+    author = {Peter Murray-Rust, Anubhab Chakraborty and others},
     year = {2021}
 }
 
