### 2022-07-05 #4451 @chartgerink
Thanks @ayush4921 - there are only two minor issues to fix still:

1. There are incomplete sentences such as the following - please proofread for reader clarity upon publication and here some of the issues I came across
  - [ ] p2
    - [ ] "See Data Some" is missing a punctuation mark
  - [ ] p5
    - [ ] "`FIRST_PDATE`" seems to miss a letter

Thanks for being so on top of this!
