### 2022-07-07 #4451 @chartgerink
Perfect - thanks @ayush4921!

There is still the missing reference DOI, would you mind adding that?

After adding it you can go ahead and tag and archive your repository on for example Zenodo or FigShare. See also https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance for the steps.

Please let me know the DOI of the archived version of your repository and then I can recommend the paper to be accepted 😊
