+++ joss/paper.bib
@@ -35,6 +35,7 @@
   author = {J., Eck N. and L, Waltman},
   date = {2010},
   title = { Software Survey: VOSviewer, a Computer Program for Bibliometric Mapping’},
+  doi = {10.1007/s11192-009-0146-3},
   volume = {84},
   language = {en},
   journal = {Scientometrics},
