+++ joss/paper.bib
@@ -54,7 +54,7 @@
     title = {pyami - Semantic Reader of the Scientific Literature},
     version = {0.0.17},
     url = {https://github.com/petermr/pyami/},
-    author = {Peter Murray-Rust and others},
+    author = {Peter Murray-Rust},
     year = {2021}
 }
 
@@ -62,7 +62,7 @@
     title = {docanalysis - Unsupervised entity extraction},
     version = {0.0.7},
     url = {https://github.com/petermr/docanalysis},
-    author = {Ayush Garg, Shweata N. Hedge and others},
+    author = {Ayush Garg, Shweata N. Hedge},
     year = {2021}
 }
 
@@ -70,7 +70,7 @@
     title = {pyamiimage - tools to extract semantic information from scientific diagrams},
     version = {0.0.8},
     url = {https://github.com/petermr/pyamiimage},
-    author = {Peter Murray-Rust, Anubhab Chakraborty and others},
+    author = {Peter Murray-Rust, Anubhab Chakraborty},
     year = {2021}
 }
 
@@ -78,7 +78,7 @@
     title = {getpapers - tools to extract semantic information from scientific diagrams},
     version = {0.4.17},
     url = {https://github.com/ContentMine/getpapers},
-    author = {Rik Smith-Unna and others},
+    author = {Rik Smith-Unna},
     year = {2021}
 }
 
