+++ joss/paper.bib
@@ -70,7 +70,7 @@
     title = {pyamiimage - tools to extract semantic information from scientific diagrams},
     version = {0.0.8},
     url = {https://github.com/petermr/pyamiimage},
-    author = {Peter Murray-Rust, Anubhab Chakraborty},
+    author = {Peter Murray-Rust, Anubhab Chakraborty, V.S.M. Roopa},
     year = {2021}
 }
 
