### 2022-07-07 #4451 @ayush4921
Thank you @khinsen @danielskatz @chartgerink @VeroniqueLegrand and @kjappelbaum for all your valuable feedback. I am happy to say that my first manuscript was with JOSS.
